fix #716, the conflicts of SlotChainBuilder SPI
