+++ spring-cloud-alibaba-sentinel-gateway/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,4 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 org.springframework.cloud.alibaba.sentinel.gateway.zuul.SentinelZuulAutoConfiguration,\
-org.springframework.cloud.alibaba.sentinel.gateway.scg.SentinelSCGAutoConfiguration
+org.springframework.cloud.alibaba.sentinel.gateway.scg.SentinelSCGAutoConfiguration,\
+org.springframework.cloud.alibaba.sentinel.gateway.SentinelGatewayAutoConfiguration
